+++ test/unittest/TSegments.cpp
@@ -99,8 +99,8 @@ TEST(MPSegmentTest, Test1) {
   ASSERT_EQ("[\"南京市\", \"长江大桥\"]", s << words);
 
   // MaxWordLen
-  //ASSERT_TRUE(segment.cut("南京市长江大桥", words, 3));
-  //ASSERT_EQ("[\"南京市\", \"长江\", \"大桥\"]", s << words);
+  ASSERT_TRUE(segment.cut("南京市长江大桥", words, 3));
+  ASSERT_EQ("[\"南京市\", \"长江\", \"大桥\"]", s << words);
 }
 
 //TEST(MPSegmentTest, Test2) {