+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # CppJieba ChangeLog
 
+## next version
+
+1. 使用工程上比较 tricky 的 Trie树优化办法。废弃了之前的 `Aho-Corasick-Automation` 实现，可读性更好，性能更高。
+
 ## v3.1.0
 
 1. 新增可动态增加词典的API: insertUserWord