+++ ChangeLog.md
@@ -3,7 +3,8 @@
 ## next version
 
 1. 使用工程上比较 tricky 的 Trie树优化办法。废弃了之前的 `Aho-Corasick-Automation` 实现，可读性更好，性能更高。
-2. 新增层次分词器: LevelSegment
+2. 新增层次分词器: LevelSegment 。
+3. 增加MPSegment的细粒度分词功能。
 
 ## v3.1.0