+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # CppJieba ChangeLog
 
-## next version
+## v3.2.0
 
 1. 使用工程上比较 tricky 的 Trie树优化办法。废弃了之前的 `Aho-Corasick-Automation` 实现，可读性更好，性能更高。
 2. 新增层次分词器: LevelSegment 。